refactor(module): product 모듈 분리
- org.flywaydb:flyway-core 모듈 디팬던시 root만 갖도록 변경
  - 해당 관리 sql을 root에만 두기 때문
- unable to find @SpringBootConfiguration 해결
  - 신규 생성한 모듈의 테스트 패키지 내 @SpringBootApplication 어노테이션이 붙은 클래스가 존재하지 않아 테스트가 제대로 작동 안함 해결
- truncate.sql 각 모듈별로 관리
  - product-api 모듈의 경우 PRODUCT 테이블만 알기 때문에 truncate도 PRODUCT만 하도록 함
- 모듈 테스트 시 타 모듈의 클래스 import 안됨 해결
  - testCompile을 build.gradle에 추가하여 해결
