+++ build.gradle
@@ -29,8 +29,6 @@ allprojects {
         implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
         implementation 'org.springframework.boot:spring-boot-starter-web'
 
-        implementation 'org.flywaydb:flyway-core'
-
         runtimeOnly 'com.h2database:h2'
 
         testImplementation 'io.rest-assured:rest-assured:3.3.0'
@@ -43,10 +41,24 @@ allprojects {
 }
 
 dependencies {
-    implementation project(':kitchenpos-common')
+    implementation 'org.flywaydb:flyway-core'
+
+    compile project(':kitchenpos-common')
+    compile project(':kitchenpos-product-api')
+
+    testCompile project(':kitchenpos-common').sourceSets.test.output
+    testCompile project(':kitchenpos-product-api').sourceSets.test.output
 }
 
 project(':kitchenpos-common') {
     dependencies {
     }
 }
+
+project(':kitchenpos-product-api') {
+    dependencies {
+        implementation project(':kitchenpos-common')
+        testCompile project(':kitchenpos-common').sourceSets.test.output
+    }
+}
+
