+++ kitchenpos-common/src/test/java/kitchenpos/common/acceptance/AcceptanceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos;
+package kitchenpos.common.acceptance;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
