+++ kitchenpos-product-api/build.gradle
@@ -1,0 +1,6 @@
+bootJar {
+    enabled = false
+}
+jar {
+    enabled = true
+}
