+++ kitchenpos-product-api/src/test/java/kitchenpos/product/ApplicationTest.java
@@ -1,0 +1,11 @@
+package kitchenpos.product;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ApplicationTest {
+    @Test
+    void contextLoads() {
+    }
+}
