+++ kitchenpos-product-api/src/test/resources/sql/truncate.sql
@@ -1,0 +1,3 @@
+SET FOREIGN_KEY_CHECKS = 0;
+TRUNCATE TABLE PRODUCT;
+SET FOREIGN_KEY_CHECKS = 1;
