+++ settings.gradle
@@ -1,3 +1,4 @@
 rootProject.name = 'kitchenpos'
 include 'kitchenpos-common'
+include 'kitchenpos-product-api'
