+++ src/test/java/kitchenpos/menugroup/acceptance/MenuGroupAcceptanceTest.java
@@ -11,7 +11,7 @@ import io.restassured.response.Response;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.AcceptanceTest;
+import kitchenpos.common.acceptance.AcceptanceTest;
 import kitchenpos.menugroup.dto.MenuGroupRequest;
 import kitchenpos.menugroup.dto.MenuGroupResponse;
 import org.junit.jupiter.api.BeforeEach;
