+++ src/test/resources/sql/truncate.sql
